Oregon Treaty Tweaks
Made an decision so CAN/USA can automatically colonize the empty Oregon territory after the conclusion of the 54-40 event chain. Can still be colonized before that but the AI will almost never do so because CAN is over their colonial points limit and the USA is never a GP/SP for long enough.